Add Employee page

+++ src/main/java/com/worldbestsoft/model/Employee.java
@@ -15,7 +15,7 @@ import javax.persistence.Table;
 @Table(name = "employee", catalog = "onion")
 public class Employee implements java.io.Serializable {
 
-	private int id;
+	private Long id;
 	private String firstName;
 	private String lastName;
 	private String nickName;
@@ -28,11 +28,11 @@ public class Employee implements java.io.Serializable {
 	public Employee() {
 	}
 
-	public Employee(int id) {
+	public Employee(Long id) {
 		this.id = id;
 	}
 
-	public Employee(int id, String firstName, String lastName, String nickName, Integer age, String idCardNo, String address, BigDecimal wage, String memo) {
+	public Employee(Long id, String firstName, String lastName, String nickName, Integer age, String idCardNo, String address, BigDecimal wage, String memo) {
 		this.id = id;
 		this.firstName = firstName;
 		this.lastName = lastName;
@@ -46,11 +46,11 @@ public class Employee implements java.io.Serializable {
 
 	@Id
 	@Column(name = "id", unique = true, nullable = false)
-	public int getId() {
+	public Long getId() {
 		return this.id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
